Move the generate_titles method into the ability as we're no longer passing the feature into the ability. Move the get_post_context method into a helper function so other functionality can take advantage of this. Move the system instructions into the ability
