+++ includes/Abilities/Title_Generation/Title_Generation.php
@@ -7,11 +7,13 @@
 
 declare( strict_types=1 );
 
-namespace WordPress\AI\Abilities;
+namespace WordPress\AI\Abilities\Title_Generation;
 
 use WP_Error;
+use WordPress\AI\API_Request;
 use WordPress\AI\Abstracts\Abstract_Ability;
 
+use function WordPress\AI\get_post_context;
 use function WordPress\AI\normalize_content;
 
 /**
@@ -20,14 +22,6 @@ use function WordPress\AI\normalize_content;
  * @since 0.1.0
  */
 class Title_Generation extends Abstract_Ability {
-
-	/**
-	 * The Feature class that the ability belongs to.
-	 *
-	 * @since 0.1.0
-	 * @var \WordPress\AI\Features\Title_Generation\Title_Generation
-	 */
-	protected $feature;
 
 	/**
 	 * Returns the input schema of the ability.
@@ -103,11 +97,6 @@ class Title_Generation extends Abstract_Ability {
 			),
 		);
 
-		// Setup the context we want to pass to the AI.
-		$context = array(
-			'content' => normalize_content( $args['content'] ?? '' ),
-		);
-
 		// If a post ID is provided, ensure the post exists before using its' content.
 		if ( $args['post_id'] ) {
 			$post = get_post( $args['post_id'] );
@@ -120,19 +109,21 @@ class Title_Generation extends Abstract_Ability {
 				);
 			}
 
-			// Default to the passed in content but fallback to the post content otherwise.
-			if ( ! $args['content'] ) {
-				$args['content'] = apply_filters( 'the_content', $post->post_content );
-				$context         = array(
-					'content' => normalize_content( $args['content'] ),
-				);
-			}
+			// Get the post context.
+			$context = get_post_context( $args['post_id'] );
 
-			$context = array_merge( $context, $this->get_context( $args['post_id'] ) );
+			// Default to the passed in content if it exists.
+			if ( $args['content'] ) {
+				$context['content'] = normalize_content( $args['content'] );
+			}
+		} else {
+			$context = array(
+				'content' => normalize_content( $args['content'] ?? '' ),
+			);
 		}
 
 		// If we have no content, return an error.
-		if ( ! $args['content'] ) {
+		if ( empty( $context['content'] ) ) {
 			return new WP_Error(
 				'content_not_provided',
 				esc_html__( 'Content is required to generate title suggestions.', 'ai' )
@@ -140,7 +131,7 @@ class Title_Generation extends Abstract_Ability {
 		}
 
 		// Generate the titles.
-		$result = $this->feature->generate_titles( $context, $args['n'] );
+		$result = $this->generate_titles( $context, $args['n'] );
 
 		// If we have an error, return it.
 		if ( is_wp_error( $result ) ) {
@@ -219,53 +210,49 @@ class Title_Generation extends Abstract_Ability {
 	}
 
 	/**
-	 * Returns the context for the given post ID.
+	 * Generates title suggestions from the given content.
 	 *
 	 * @since 0.1.0
 	 *
-	 * @param int $post_id The ID of the post to get the context for.
-	 * @return array<string, string> The context for the given post ID.
+	 * @param string|array<string, string> $context The context to generate a title from.
+	 * @param int $n The number of titles to generate.
+	 * @return array<string>|\WP_Error The generated titles, or a WP_Error if there was an error.
 	 */
-	protected function get_context( int $post_id ): array {
-		$post    = get_post( $post_id );
-		$context = array();
-
-		// If the post doesn't exist, return early.
-		if ( ! $post ) {
-			return $context;
+	protected function generate_titles( $context, int $n = 1 ) {
+		// Convert the context to a string if it's an array.
+		if ( is_array( $context ) ) {
+			$context = implode(
+				"\n",
+				array_map(
+					static function ( $key, $value ) {
+						return sprintf(
+							'%s: %s',
+							ucwords( str_replace( '_', ' ', $key ) ),
+							$value
+						);
+					},
+					array_keys( $context ),
+					$context
+				)
+			);
 		}
 
-		if ( $post->post_title ) {
-			$context['current_title'] = $post->post_title;
+		// Make our request.
+		$request  = new API_Request();
+		$response = $request->generate_text(
+			'"""' . $context . '"""',
+			$this->get_system_instruction(),
+			array(
+				'candidateCount' => (int) $n,
+				'temperature'    => 0.7,
+			)
+		);
+
+		// If we have an error, return it.
+		if ( is_wp_error( $response ) ) {
+			return $response;
 		}
 
-		if ( $post->post_name ) {
-			$context['slug'] = $post->post_name;
-		}
-
-		$author = get_user_by( 'ID', $post->post_author );
-		if ( $author ) {
-			$context['author'] = $author->display_name;
-		}
-
-		if ( $post->post_type ) {
-			$context['content_type'] = $post->post_type;
-		}
-
-		if ( $post->post_excerpt ) {
-			$context['excerpt'] = $post->post_excerpt;
-		}
-
-		$categories = get_the_terms( $post_id, 'category' );
-		if ( $categories && ! is_wp_error( $categories ) ) {
-			$context['categories'] = implode( ', ', wp_list_pluck( $categories, 'name' ) );
-		}
-
-		$tags = get_the_terms( $post_id, 'post_tag' );
-		if ( $tags && ! is_wp_error( $tags ) ) {
-			$context['tags'] = implode( ', ', wp_list_pluck( $tags, 'name' ) );
-		}
-
-		return $context;
+		return $response;
 	}
 }
