+++ includes/Features/Title_Generation/Title_Generation.php
@@ -9,8 +9,7 @@ declare( strict_types=1 );
 
 namespace WordPress\AI\Features\Title_Generation;
 
-use WordPress\AI\API_Request;
-use WordPress\AI\Abilities\Title_Generation as Title_Generation_Ability;
+use WordPress\AI\Abilities\Title_Generation\Title_Generation as Title_Generation_Ability;
 use WordPress\AI\Abstracts\Abstract_Feature;
 
 /**
@@ -58,52 +57,5 @@ class Title_Generation extends Abstract_Feature {
 				'ability_class' => Title_Generation_Ability::class,
 			),
 		);
-	}
-
-	/**
-	 * Generates title suggestions from the given content.
-	 *
-	 * @since 0.1.0
-	 *
-	 * @param string|array<string, string> $context The context to generate a title from.
-	 * @param int $n The number of titles to generate.
-	 * @return array<string>|\WP_Error The generated titles, or a WP_Error if there was an error.
-	 */
-	public function generate_titles( $context, int $n = 1 ) {
-		// Convert the context to a string if it's an array.
-		if ( is_array( $context ) ) {
-			$context = implode(
-				"\n",
-				array_map(
-					static function ( $key, $value ) {
-						return sprintf(
-							'%s: %s',
-							ucwords( str_replace( '_', ' ', $key ) ),
-							$value
-						);
-					},
-					array_keys( $context ),
-					$context
-				)
-			);
-		}
-
-		// Make our request.
-		$request  = new API_Request();
-		$response = $request->generate_text(
-			'"""' . $context . '"""',
-			$this->get_system_instruction(),
-			array(
-				'candidateCount' => (int) $n,
-				'temperature'    => 0.7,
-			)
-		);
-
-		// If we have an error, return it.
-		if ( is_wp_error( $response ) ) {
-			return $response;
-		}
-
-		return $response;
 	}
 }
